abstract away ui code from operator
